+++ packages/fields/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @keystone-alpha/fields
 
+## 10.3.0
+
+### Minor Changes
+
+- [a640471a](https://github.com/keystonejs/keystone-5/commit/a640471a): - Add `Slug.alwaysMakeUnique` option to force calling `makeUnique` even when initially generated slug may already be unique to prevent accidental data leak.
+  - Fix a bug where items restricted via access control weren't considered when testing a `Slug` for uniqueness.
+
 ## 10.2.0
 
 ### Minor Changes